First prototype of an interactive alignment type.
Code name is "Equalizer".
You select a source and target concept scheme. For both a columnbrowser is shown. At the moment you can only manually align the concepts. Select a concept from the source and one from the target and click align.

The idea is to include matching algorithms to create candidate alignments. The task of the user is than to confirm the correct alignments.

Issues are:
- the visualization of aligned concepts
- the contextual information shown for each concept
- the interaction method to confirm alignments