Switch back to JavaScript module approach
I tried several methods to generate JavaScript from prolog, but did not find a satisfactory solution. I now turned to the complete opposite: generate almost no JS in prolog. I only include the YUI3 initialization and a single object that holds all the server dependent variables that are required in the JS.